Expose CHAMD as a CHAT-reading library. Resolves #4
