+++ chamd/__init__.py
@@ -1,1 +1,2 @@
 #!/usr/bin/env python3
+from .chat_reader import ChatReader
